+++ lib/src/core/managers/group_manager.dart
@@ -6,28 +6,28 @@ part of '../../../dartemis.dart';
 ///
 /// An [int] can only belong to several groups (0,n) at a time.
 class GroupManager extends Manager {
-  final Map<String, Bag<int>> _entitiesByGroup;
-  final Map<int, Bag<String>> _groupsByEntity;
+  final Map<String, EntityBag> _entitiesByGroup;
+  final Map<Entity, Bag<String>> _groupsByEntity;
 
   /// Creates the [GroupManager].
   GroupManager()
-      : _entitiesByGroup = <String, Bag<int>>{},
-        _groupsByEntity = <int, Bag<String>>{};
+      : _entitiesByGroup = <String, EntityBag>{},
+        _groupsByEntity = <Entity, Bag<String>>{};
 
   /// Set the group of the entity.
-  void add(int entity, String group) {
-    _entitiesByGroup.putIfAbsent(group, Bag<int>.new).add(entity);
+  void add(Entity entity, String group) {
+    _entitiesByGroup.putIfAbsent(group, EntityBag.new).add(entity);
     _groupsByEntity.putIfAbsent(entity, Bag<String>.new).add(group);
   }
 
   /// Remove the entity from the specified group.
-  void remove(int entity, String group) {
+  void remove(Entity entity, String group) {
     _entitiesByGroup[group]?.remove(entity);
     _groupsByEntity[entity]?.remove(group);
   }
 
   /// Remove [entity] from all existing groups.
-  void removeFromAllGroups(int entity) {
+  void removeFromAllGroups(Entity entity) {
     final groups = _groupsByEntity[entity];
     if (groups != null) {
       groups
@@ -39,21 +39,21 @@ class GroupManager extends Manager {
   }
 
   /// Get all entities that belong to the provided group.
-  Iterable<int> getEntities(String group) =>
-      _entitiesByGroup.putIfAbsent(group, Bag<int>.new);
+  Iterable<Entity> getEntities(String group) =>
+      _entitiesByGroup.putIfAbsent(group, EntityBag.new);
 
   /// Returns the groups the entity belongs to, null if none.
-  Iterable<String>? getGroups(int entity) => _groupsByEntity[entity];
+  Iterable<String>? getGroups(Entity entity) => _groupsByEntity[entity];
 
   /// Checks if the entity belongs to any group.
-  bool isInAnyGroup(int entity) => getGroups(entity) != null;
+  bool isInAnyGroup(Entity entity) => getGroups(entity) != null;
 
   /// Check if the entity is in the supplied group.
-  bool isInGroup(int entity, String group) {
+  bool isInGroup(Entity entity, String group) {
     final groups = _groupsByEntity[entity];
     return (groups != null) && groups.contains(group);
   }
 
   @override
-  void deleted(int entity) => removeFromAllGroups(entity);
+  void deleted(Entity entity) => removeFromAllGroups(entity);
 }