+++ lib/src/core/managers/player_manager.dart
@@ -4,26 +4,26 @@ part of '../../../dartemis.dart';
 ///
 /// An entity can only belong to a single player at a time.
 class PlayerManager extends Manager {
-  final Map<int, String?> _playerByEntity;
-  final Map<String, Bag<int>> _entitiesByPlayer;
+  final Map<Entity, String?> _playerByEntity;
+  final Map<String, EntityBag> _entitiesByPlayer;
 
   /// Creates the [PlayerManager].
   PlayerManager()
-      : _playerByEntity = <int, String>{},
-        _entitiesByPlayer = <String, Bag<int>>{};
+      : _playerByEntity = <Entity, String>{},
+        _entitiesByPlayer = <String, EntityBag>{};
 
   /// Make [entity] belong to [player].
-  void setPlayer(int entity, String player) {
+  void setPlayer(Entity entity, String player) {
     _playerByEntity[entity] = player;
-    _entitiesByPlayer.putIfAbsent(player, Bag<int>.new).add(entity);
+    _entitiesByPlayer.putIfAbsent(player, EntityBag.new).add(entity);
   }
 
   /// Returns all entities that belong to [player].
-  Iterable<int> getEntitiesOfPlayer(String player) =>
-      _entitiesByPlayer[player] ??= Bag<int>();
+  Iterable<Entity> getEntitiesOfPlayer(String player) =>
+      _entitiesByPlayer[player] ??= EntityBag();
 
   /// Removes [entity] from the player it is associated with.
-  void removeFromPlayer(int entity) {
+  void removeFromPlayer(Entity entity) {
     final player = _playerByEntity[entity];
     if (player != null) {
       _entitiesByPlayer[player]?.remove(entity);
@@ -31,8 +31,8 @@ class PlayerManager extends Manager {
   }
 
   /// Returns the player associated with [entity].
-  String? getPlayer(int entity) => _playerByEntity[entity];
+  String? getPlayer(Entity entity) => _playerByEntity[entity];
 
   @override
-  void deleted(int entity) => removeFromPlayer(entity);
+  void deleted(Entity entity) => removeFromPlayer(entity);
 }