+++ lib/src/core/managers/tag_manager.dart
@@ -4,16 +4,16 @@ part of '../../../dartemis.dart';
 /// entities such as "PLAYER", "BOSS" or something that is very unique.
 /// An entity can only belong to one tag (0,1) at a time.
 class TagManager extends Manager {
-  final Map<String, int?> _entitiesByTag;
-  final Map<int, String?> _tagsByEntity;
+  final Map<String, Entity?> _entitiesByTag;
+  final Map<Entity, String?> _tagsByEntity;
 
   /// Create the [TagManager].
   TagManager()
-      : _entitiesByTag = <String, int>{},
-        _tagsByEntity = <int, String>{};
+      : _entitiesByTag = <String, Entity>{},
+        _tagsByEntity = <Entity, String>{};
 
   /// Register a [tag] to an [entity].
-  void register(int entity, String tag) {
+  void register(Entity entity, String tag) {
     unregister(tag);
     _entitiesByTag[tag] = entity;
     _tagsByEntity[entity] = tag;
@@ -28,17 +28,17 @@ class TagManager extends Manager {
   bool isRegistered(String tag) => _entitiesByTag.containsKey(tag);
 
   /// Returns the entity with [tag].
-  int? getEntity(String tag) => _entitiesByTag[tag];
+  Entity? getEntity(String tag) => _entitiesByTag[tag];
 
   /// Returns the tag of the [entity].
-  String? getTag(int entity) => _tagsByEntity[entity];
+  String? getTag(Entity entity) => _tagsByEntity[entity];
 
   /// Returns all known tags.
   Iterable<String> getRegisteredTags() =>
       _tagsByEntity.values as Iterable<String>;
 
   @override
-  void deleted(int entity) {
+  void deleted(Entity entity) {
     final removedTag = _tagsByEntity.remove(entity);
     if (removedTag != null) {
       _entitiesByTag.remove(removedTag);