+++ lib/src/core/utils/entity_bag.dart
@@ -1,32 +1,31 @@
 part of '../../../dartemis.dart';
 
 /// A [Bag] that uses a [BitSet] to manage entities. Results in faster
-/// removement of entities.
-// ignore: prefer_mixin
-class EntityBag with IterableMixin<int> {
+/// removal of entities.
+class EntityBag with Iterable<Entity> {
   BitSet _entities;
 
   /// Creates an [EntityBag].
   EntityBag() : _entities = BitSet(32);
 
-  /// Add a new [element]. If the element already exists, nothing changes.
-  void add(int element) {
-    if (element >= _entities.length) {
-      _entities = BitSet.fromBitSet(_entities, length: element);
+  /// Add a new [entity]. If the entity already exists, nothing changes.
+  void add(Entity entity) {
+    if (entity._id >= _entities.length) {
+      _entities = BitSet.fromBitSet(_entities, length: entity._id + 1);
     }
-    _entities[element] = true;
+    _entities[entity._id] = true;
   }
 
-  /// Removes [element]. Returns `true` if there was an element and `false`
+  /// Removes [entity]. Returns `true` if there was an element and `false`
   /// otherwise.
-  bool remove(int element) {
-    final result = _entities[element];
-    _entities[element] = false;
+  bool remove(Entity entity) {
+    final result = _entities[entity._id];
+    _entities[entity._id] = false;
     return result;
   }
 
   @override
-  bool contains(covariant int element) => _entities[element];
+  bool contains(covariant Entity element) => _entities[element._id];
 
   @override
   int get length => _entities.cardinality;
@@ -35,5 +34,6 @@ class EntityBag with IterableMixin<int> {
   void clear() => _entities.clearAll();
 
   @override
-  Iterator<int> get iterator => _entities.toIntValues().iterator;
+  Iterator<Entity> get iterator =>
+      _entities.toIntValues().map(Entity._).iterator;
 }